Implements recent projects menu (first draft)
Also stores a "lastDirectory" in QSettings.

This is work in progress - something odd is going on with project
loading on one of my systems, though these seem to occur for any
revision of the codebase, so I don't think it's related to my "recent
projects" changes.

Need to implement a "last directory for import networks" etc.