Improve stylesheets. Fixes Github issue#8
This improves the toolbar stylesheet, which is set in the UI form
mainwindow.ui, so that it applies to certain class types, but not all
class types. This way, tooltips maintain their default look (black on
light yellow), which looks good (it's the same a QT Creator, for
example).

Similar changes to other stylesheet snippets have been made in the
code - wrapping with QLabel{} or QFrame{} etc.