Fixes a bug with python connection lists
This clears the ParameterData indices before adding new indices.

This showed up as being a problem for connection lists where changing
a parameter caused a change in the number of connections (e.g. our
Gaussian Kernel for the Oculomotor model).

This trivial fix took an entire day, as I had to satisfy myself the
python script was working, and it took all day to implement some sort
of access to the pythong script stdout!