Applies Alex Copes changes to bring metadata into the model.xml file
Instead of storing metadata about different parts of the model in a
separate file, that information is now stored within model.xml within
LL:Annotation nodes.

This change retains the code's ability to read the old format, where
the annotations are stored in a metadata.xml file.

The SpineCreator-specific information currently stored within an
Annotation element is enclosed within a SpineCreator element.