Adds AdditionalFiles code
Adds code to load a list of optional 'additional files' from a
specification in the .proj file and then copy those additional files
when saving the project.

This is useful for me in my ocularmotor model, which has a
luminances.json file which I need to be copied into the output
directory when running the model with SpineCreator.